feat: make lucene filter generic
