+++ src/adapters/kysely/users/list.ts
@@ -1,76 +1,10 @@
 import { ListUsersResponse } from "../../interfaces/Users";
 import { Database } from "../../../types";
-import { Kysely, SelectQueryBuilder } from "kysely";
+import { Kysely } from "kysely";
 import { ListParams } from "../../interfaces/ListParams";
 import getCountAsInt from "../../../utils/getCountAsInt";
 import { parseBooleans } from "./booleans";
-
-const searchableColumns = ["email", "name"];
-
-// This could probably be made generic
-function luceneFilter(
-  db: Kysely<Database>,
-  qb: SelectQueryBuilder<Database, "users", {}>,
-  query: string,
-) {
-  const filters = query.split(/\s+/).map((filter) => {
-    let isNegation = filter.startsWith("-");
-    let key, value, isExistsQuery;
-
-    if (filter.startsWith("-_exists_:")) {
-      key = filter.substring(10); // Remove '-_exists_:' part
-      isExistsQuery = true;
-      isNegation = true;
-    } else if (filter.startsWith("_exists_:")) {
-      key = filter.substring(9); // Remove '_exists_:' part
-      isExistsQuery = true;
-      isNegation = false;
-    } else if (filter.includes(":")) {
-      isNegation = filter.startsWith("-");
-      [key, value] = isNegation
-        ? filter.substring(1).split(":")
-        : filter.split(":");
-      isExistsQuery = false;
-    } else {
-      // Single word search case
-      key = null;
-      value = filter;
-      isExistsQuery = false;
-    }
-
-    return { key, value, isNegation, isExistsQuery };
-  });
-
-  // Apply filters to the query builder
-  filters.forEach(({ key, value, isNegation, isExistsQuery }) => {
-    if (key) {
-      if (isExistsQuery) {
-        if (isNegation) {
-          // I'm not following how this ever worked...
-          qb = qb.where(key as any, "is", null);
-        } else {
-          qb = qb.where(key as any, "is not", null);
-        }
-      } else {
-        if (isNegation) {
-          qb = qb.where(key as any, "!=", value);
-        } else {
-          qb = qb.where(key as any, "=", value);
-        }
-      }
-    } else {
-      const { ref } = db.dynamic;
-      // Generic single-word search across specified columns
-      qb = qb.where((eb) =>
-        eb.or(
-          searchableColumns.map((col) => eb(ref(col), "like", `%${value}%`)),
-        ),
-      );
-    }
-  });
-
-  return qb;
-}
+import { luceneFilter } from "../helpers/filter";
 
 export function listUsers(db: Kysely<Database>) {
   return async (
@@ -79,7 +13,7 @@ export function listUsers(db: Kysely<Database>) {
   ): Promise<ListUsersResponse> => {
     let query = db.selectFrom("users").where("users.tenant_id", "=", tenantId);
     if (params.q) {
-      query = luceneFilter(db, query, params.q);
+      query = luceneFilter(db, query, params.q, ["email", "name"]);
     }
 
     if (params.sort && params.sort.sort_by) {
